services/api-gateway/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
boto3==1.34.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.3